+++ CHANGES.markdown
@@ -1,3 +1,8 @@
+## Changes in 1.6.0
+ - Rename hspecWith to hspecWithResult
+ - Add a new hspecWith, which interprets command-line arguments and exits with
+   exitFailure if there is a failing spec
+
 ## Changes in 1.5.4
  - Make sure that QuickCheck is never chatty
  - Make sure progress for QuickCheck examples is shown